LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/timeline-range-name-offset-in-keyframes.tentative-expected.txt


FAIL Timeline offset in Animation Keyframes assert_equals: Effect at entry 50% expected "0.5" but got "0.2375"
